semi implementation of arrays + objects and convert the runtime from being stack based to register based

+++ src/shared/environment.cpp
@@ -9,14 +9,14 @@ void Environment::add_host_func(const std::string& name, host_function func) {
     host_funcs.push_back(func);
 }
 
-std::optional<uint64_t> Environment::get_func_id(const std::string& name) {
+std::optional<uint16_t> Environment::get_func_id(const std::string& name) {
     if (!name_mapping.contains(name)) {
         return std::nullopt;
     }
     return name_mapping[name];
 }
 
-void Environment::invoke_function(runtime::Runtime* rt, uint64_t id, uint64_t nargs){
+void Environment::invoke_function(runtime::Runtime* rt, uint16_t id, uint8_t nargs){
     host_funcs[id](rt, nargs);
 }
