+++ src/shared/environment.h
@@ -20,13 +20,13 @@ typedef void (*host_function) (runtime::Runtime* rt, uint64_t nargs);
 class Environment {
 public:
     void add_host_func(const std::string& name, host_function func);
-    std::optional<uint64_t> get_func_id(const std::string& name);
-    void invoke_function(runtime::Runtime* rt, uint64_t id, uint64_t nargs);
+    std::optional<uint16_t> get_func_id(const std::string& name);
+    void invoke_function(runtime::Runtime* rt, uint16_t id, uint8_t nargs);
 
     runtime::Heap heap;
 private:
 
-    std::unordered_map<std::string, uint64_t> name_mapping;
+    std::unordered_map<std::string, uint16_t> name_mapping;
     std::vector<host_function> host_funcs;
 };
